requirements.txt
asgiref==3.7.2
Django==5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.52.0
dj-rest-auth[with_social]==4.0.0
django-cors-headers==4.1.0
psycopg2==2.9.9
python-dotenv==1.0.0
whitenoise==6.6.0
requests-oauthlib==1.3.1
opencensus-ext-azure==1.1.12
PyJWT==2.8.0
requests==2.31.0
channels[daphne]==4.0.0
channels-redis==4.1.0
python-socketio[client]==5.10.0
azure-storage-blob==12.19.0
beautifulsoup4==4.12.2
django-filter==23.5
django-storages[azure]==1.14.2
pillow==10.2.0
django-select2==8.1.2
django-json-widget==2.0.1
stripe==9.8.0
azure-communication-email==1.0.0
exa-py==1.0.16
